Batched data loading and visualization updates
* rename objectron datasets for disambiguation
* parallel loading support in detection objectron
* fix cached dataset loading in keypoint regression
* fix tensor shape inconsistencies regarding instance num
* use gcs + tfrecord instead of torch_xla everywhere
* add missing feature keys to objectron dataset
* add configurable instance-padding transform
* add collation loading test script
* deprecate gcs arg on tfrecord loaders wrt upstream change
* split transforms into multiple files
* modify number of keypoints computed from displacement map
* refactor out common data loader
* fix test_collate to use default loader
* cleanup keypoint regression imports
* expose transform settings to app settings
* add displacement map visualization
* fix device allocation and add loader opts
